# Speed up registerizeHarder variable conflict checking