Results in a 50% speedup in sqlite compilation!

This PR is based on the observation that vector access is much faster than set/map access. By mapping names to numbers and using a vector we can speed up all conflict setting (first 3 commits). I then mitigate the perf hit to iteration by filtering the vector outside of the loop which would iterate over it (final commit).

All asm3 tests pass, as does test_other. I should probably be noting that I don't have spidermonkey installed so some parts of tests (e.g. actual sqlite execution, though I did check that with nodejs) aren't being run. This obviously also affects the before/after for sqlite below.

Before (after #3368):
```
~/em/emsdk/emscripten/incoming/tests $ python runner.py asm3.test_sqlite 2>&1 | grep 'test in'
Ran 1 test in 61.199s
~/em/emsdk/emscripten/incoming/tests $ EMCC_DEBUG=1 emcc -O3 --llvm-lto 1 -s EMULATE_FUNCTION_POINTER_CASTS=1 -o out.js python/python.bc 2>&1 | grep 'js opts'
DEBUG    root: emcc step "js opts" took 5.23 seconds
~/em/emsdk/emscripten/incoming/tests $ cd ~/em/BananaBread/cube2/src/web
~/em/BananaBread/cube2/src/web $ (make && time make client server) 2>&1 | grep real
real    0m28.940s
```

After this pr:
```
~/em/emsdk/emscripten/incoming/tests $ python runner.py asm3.test_sqlite 2>&1 | grep 'test in'
Ran 1 test in 29.944s
~/em/emsdk/emscripten/incoming/tests $ EMCC_DEBUG=1 emcc -O3 --llvm-lto 1 -s EMULATE_FUNCTION_POINTER_CASTS=1 -o out.js python/python.bc 2>&1 | grep 'js opts'
DEBUG    root: emcc step "js opts" took 3.17 seconds
~/em/emsdk/emscripten/incoming/tests $ cd ~/em/BananaBread/cube2/src/web
~/em/BananaBread/cube2/src/web $ (make && time make client server) 2>&1 | grep real
real    0m18.117s
```