@kripken 2015-04-21
I haven't read the patches yet, but the js opts part on sqlite goes from 35.84 seconds to... 5.73 (!). That's much more than a 50% speedup! But it's so big I was kind of worried that something changed in the output for the worse, so I'll ran the benchmark suite on this... and it looks like code speed and size remain the same. Awesome!

Starting to read the patches now.