@aidanhs 2015-04-21
Yes, my 'flagship measurement' at the top is based on the *overall* compilation time, just as an interesting demonstration of how much js optimisation was previously dominating compilation time. Bananabread is similarly measuring overall compilation time. Only python drills down to js opts time. If I start putting together more optimisations maybe I'll try and present both the individual and overall speedup for clarity. Though maybe I need to focus elsewhere now :)

I too was very surprised, which is why I ran all of test_other and asm3 *and* tweaked the sqlite test to run under nodejs on my PC just to check it actually ran.