@aidanhs 2015-04-21
FYI, I've just pushed a tiny followup commit to improve code clarity. It has no performance impact I can measure (possibly a very slight improvement).