@kripken 2015-04-21
Ok, that addressed all my questions, thanks. Merging.